Add support for actions based on Docker images
Previously, client didn't provide any way to update an action which
was based on a Docker image, since it didn't provide any way to
specify the target image. This was required for the API endpoint to
accept a request.

This commit fixes it by enabling to specify the Docker image, and thus
to work with such actions.