Add support for timeout limit parameter
Previously, it was impossible to use this client to update or
create an action with a different timeout than the default.
Worse than that, using this client meant that timeout was
reset to default each time an action was updated.

This commit fixes this limitation by adding support for
custom timeout.